requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
mediapipe==0.10.21
numpy==1.26.4
pillow==10.4.0